Try a different approach to importing "cloudflare:workers".
Fixes #23

Per #23, many bundlers do not like our dynamic import of "cloudflare:workers". This PR takes a different approach, using the "exports" feature of "package.json" to point workerd at a different version of the code. That version imports "cloudflare:workers" and sticks it in the global scope before importing the rest of the library, thus allowing it to conditionally probe.