Grab bag of issue fixes! (#105)
* Add changeset for past PR #82.

* Throw an error when attempting to access RpcTarget instance properties.

This helps people learn why instance properties are not accessible over RPC, whereas returning `undefined` leaves them confused.

Fixes #55

* Implement toString() for RpcStub and RpcPromise.

They just return `[object RpcStub]` and `[object RpcPromise]`, but that's better than the previous `[object Function]` which was confusing.

As suggested in #55.

* Fix type signature of newHttpBatchRpcSession().

An earlier version of the function was designed to have the same signature as `fetch()`, but when I added `RpcSessionOptions` it made more sense to make that the second param... but I only updated the implementation and forgot to update the type.

Fixes #67.

* Support serializing Infinity, -Infinity, and NaN.

Fixes #80

* Polyfill Promise.withResolvers().

This hopefully improves compatibility with old Safari versions and Hermes (React Native).

Unfortunately it didn't seem easy to extend the tests to cover React Native. There is a vitest-react-native package, but it looks little-used and unmaintained, so I didn't want to install it.

Might help with #91, though I won't declare that one "fixed" until we actually have tests proving it.

* Document missing expression types in protocol.md.

Fixes #48